Experimental

+++ includes/ray.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ray.h                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: daspring <daspring@student.42heilbronn.    +#+  +:+       +#+        */
+/*   By: llacsivy <llacsivy@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 15:16:05 by daspring          #+#    #+#             */
-/*   Updated: 2024/11/26 13:44:22 by daspring         ###   ########.fr       */
+/*   Updated: 2024/12/06 13:19:59 by llacsivy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,8 +26,7 @@ typedef struct s_ray
 	t_tuple	direction_vec;
 }		t_ray;
 
-// t_ray	create_ray(t_tuple E, t_tuple P);
 t_ray	create_ray(size_t x_pixel, size_t y_pixel);
-t_tuple	ray_at_t(t_ray *ray, float t);
+t_tuple	ray_at_t(t_ray *ray, double t);
 
 #endif
